update_mtime(): Write update time to CGit agefile as last recourse
Adjust the match arms to remove a bit of indentation.

Add some 'anyhow' context to errors for better error reporting.

When "repo/refs/heads/[default-branch]" or "repo/packed-refs" files
don't exist, create a "repo/info/web/last-modified" file and set this
file as the CGit agefile in the repo's local 'cgitrc' file.

It's possible for a repo to not have either of the first two files when
the repo is empty and has no commits.