git::mirror(): Change HEAD to GitHub default branch
The default branch after mirroring was typically 'master'. On GitHub,
the default branch may not necessarily be 'master'. Change the default
branch by changing the HEAD to GitHub's default branch so that the
mirrored repository better matches GitHub.

We'll also need to make a change to the update function in case the
default branch changes after mirroring.