github::Repo: Change `git_url` to `clone_url`
I was getting errors mirroring and updating:

    failed to connect to github.com: Connection timed out; class=Os (2)

and

    remote 'origin' already exists; class=Config (7); code=Exists(-4)

It turns out that the `git_url` field, which I had been using previously
to mirror and clone repositories, stopped working. My guess is that it's
because Reflectub is not authorised to clone GitHub "git://" URLs, so
the connection timed out. I'm not sure why this stopped being allowed,
though.

The URL change seems to have happened around March 2022, or at least
between December 2021 and April 2022.

The second error was caused by a previously-created repository existing
in the filesystem, but not being in the database as it hadn't been
correctly mirrored. For now, I've decided not to fix that problem and am
only fixing the URL issue.

The GitHub API also includes a `clone_url` field, which contains an
HTTPS clone URL. Using this URL to mirror fixes the timeout problem.