database::Repo: From<&github::Repo>: Use newest update date
It turns out the GitHub `updated_at` field doesn't change when new
commits are pushed to the repository, only when the repository config
etc. changes.

In order to update the mirrors when any update happens in the
repository, we need to look at both of those date values to see if
they've been updated.

Take the most recent of `updated_at` or `pushed_at` and set it to
`(database::Repo).updated_at`. This allows us to refresh the repo when
either of those dates change, catching all GitHub repo updates.